requirements.txt
apscheduler>=3.10.4
moviepy==1.0.3
numpy>=1.26.0
openai>=1.6.0
Pillow>=10.0.0
pytz>=2023.3
python-dotenv>=1.0.0
requests>=2.31.0
imageio-ffmpeg>=0.4.9
